Virtual B Trees ( LRU )
This is a very simple implementation of LRU Algorithm where the least recently used pages are thrown out of the cache and the page which is just referred is put on a top priority.
I have made the implementation using stacks. The implementation can be done using Linked Lists as well.